Add clear endtime/duration button; Add lower bound to heatmap domain

+++ web/src/routes/admin/Table.svelte
@@ -1,12 +1,17 @@
 <script lang="ts">
-	import { ChevronLeft, ChevronRight, Trash } from 'lucide-svelte';
+	import { ChevronLeft, ChevronRight, Eraser, RefreshCw, Trash } from 'lucide-svelte';
 
 	import TimeSinceDisplay from '$lib/TimeSinceDisplay.svelte';
 	import * as Utils from '$lib/utils';
 	import TableButton from './TableButton.svelte';
 
+	let confirmDeleteId: string = $state('');
+	let confirmClearEndTimeId: string = $state('');
+
 	interface Props {
 		metrics: Utils.metric[];
+		onRefresh: () => Promise<void>;
+		onClearEndTimeAndDuration: (id: string) => Promise<void>;
 		onDelete: (id: string) => Promise<void>;
 		onPageChange: (changeTo: number) => Promise<void>;
 		totalItems: number;
@@ -18,6 +23,8 @@
 
 	let {
 		metrics,
+		onRefresh,
+		onClearEndTimeAndDuration,
 		onDelete,
 		onPageChange,
 		totalItems,
@@ -31,7 +38,12 @@
 <div class="flex flex-col items-center justify-center">
 	<table class="size-full table-auto overflow-scroll border border-gray-400">
 		<caption
-			>Metrics Management <i>(viewing {itemRangeLower}-{itemRangeHigher} out of {totalItems})</i
+			><b>Metrics Management</b>
+			<i>(viewing {itemRangeLower}-{itemRangeHigher} out of {totalItems})</i><button
+				class="transition-transform hover:rotate-180"
+				onclick={async () => {
+					onRefresh();
+				}}><RefreshCw /></button
 			></caption
 		>
 		<tbody>
@@ -47,14 +59,46 @@
 				<tr class="border border-gray-400 even:bg-gray-100">
 					<td>{metric.id}</td>
 					<td>{metric.startTime.toLocaleString()}</td>
-					<td>{metric.endTime.toLocaleString()}</td>
-					<td><TimeSinceDisplay timeSince={metric.duration} showMilliseconds /></td>
+					<td
+						><span
+							>{metric.endTime.toLocaleString()}{#if confirmClearEndTimeId === metric.id}
+								<button
+									class="text-red-500"
+									onclick={async () => {
+										onClearEndTimeAndDuration(metric.id);
+										confirmClearEndTimeId = '';
+									}}><span><Eraser /> Confirm Clear End Time And Duration</span></button
+								>
+							{:else}
+								<button
+									onclick={() => {
+										confirmClearEndTimeId = metric.id;
+									}}><Eraser /></button
+								>
+							{/if}</span
+						></td
+					>
+					<td
+						><TimeSinceDisplay
+							timeSince={Utils.nanosecondsToTimeSince(metric.duration)}
+							showMilliseconds
+						/></td
+					>
 					<td>
-						<button
-							onclick={async () => {
-								onDelete(metric.id);
-							}}><Trash /></button
-						>
+						{#if confirmDeleteId === metric.id}
+							<button
+								class="text-red-500"
+								onclick={async () => {
+									onDelete(metric.id);
+								}}><span><Trash /> Confirm Delete</span></button
+							>
+						{:else}
+							<button
+								onclick={() => {
+									confirmDeleteId = metric.id;
+								}}><Trash /></button
+							>
+						{/if}
 					</td>
 				</tr>
 			{/each}
